requirements.txt
anthropic>=0.18.1
pyzotero>=1.5.18
python-dotenv>=1.0.1
PyMuPDF>=1.23.26
docling>=0.1.0
pytesseract>=0.3.10
llama-cpp-python>=0.2.56 
transformers
torch